+++ Title.h
@@ -13,7 +13,7 @@ class Title : public IScene
 {
 public:
 	void update(float deltatime) override;
-	void draw(uint64_t deltatime) override;
+	void draw(float deltatime) override;
 	void init() override;
 	void loadAsync() override;	
 	void dispose() override;
