+++ system/IScene.h
@@ -8,7 +8,7 @@ public:
 	virtual ~IScene() = default;
 	virtual void update(float delta) = 0;
 	//virtual void update(float delta) = 0;
-	virtual void draw(uint64_t delta) = 0;
+	virtual void draw(float delta) = 0;
 	virtual void init() = 0;
 	virtual void loadAsync()=0;
 	virtual void dispose() = 0;
